use /data as mounted volume; updated dev doc
mounting one directory for the RDF data allows both the db file and
rdf store directory to be available. previously the db file mounted
as a directory which resulted in an empty directory.

this explains the experience with empty namespace prefixes since the
db file did not contain any namespaces.